Fix styleguide and variable type consistency

+++ project-examples/ethereum/contracts/Design Patterns/Mutex Pattern/MutexAntipattern.sol
@@ -9,12 +9,11 @@ contract MutexAntipattern {
         balances[msg.sender] += msg.value;
     }
 
-    function withdraw(uint _amount) external {
+    function withdraw(uint256 _amount) external {
         require(balances[msg.sender] >= _amount, "No balance to withdraw.");
         // This is wrong according to the Checks-Effects-Interaction Pattern.
         // For demonstration purposes only!
         msg.sender.call{value: _amount}("");
-        balances[msg.sender] -= _amount;
-        
+        balances[msg.sender] -= _amount;        
     }
 }
